Add files via upload
This was one of my first projects with the TripleTen program. This project was aimed to analyze the operations of a "superstore" to increase profitability and avoid bankruptcy. The analysis is conducted through various visualizations created in Tableau, focusing on different aspects of the store's performance. 1. Profits & Losses 2. Advertising 3. Returned Items.